@@ -5,14 +5,17 @@ apply plugin: 'kotlin-kapt'
 apply plugin: 'com.google.gms.google-services'
 apply plugin: 'com.google.firebase.crashlytics'
 
+def localProperties = new Properties()
+localProperties.load(new FileInputStream(rootProject.file("local.properties")))
+
 android {
-    compileSdkVersion 31
+    compileSdkVersion 32
     defaultConfig {
         applicationId "knf.kuma"
         minSdkVersion 21
-        targetSdkVersion 31
-        versionCode 193
-        versionName "5.1.24"
+        targetSdkVersion 32
+        versionCode 196
+        versionName "5.1.27"
         multiDexEnabled true
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         javaCompileOptions {
