@@ -1,8 +1,13 @@
 package knf.kuma
 
+
+import android.util.Log
 import androidx.test.InstrumentationRegistry
 import androidx.test.runner.AndroidJUnit4
-import org.junit.Assert.assertEquals
+import androidx.work.Configuration
+import androidx.work.testing.SynchronousExecutor
+import androidx.work.testing.WorkManagerTestInitHelper
+import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -13,12 +18,19 @@ import org.junit.runner.RunWith
  */
 @RunWith(AndroidJUnit4::class)
 class ExampleInstrumentedTest {
+    @Before
+    fun setup() {
+        val context = InstrumentationRegistry.getTargetContext()
+        val config = Configuration.Builder()
+            .setMinimumLoggingLevel(Log.DEBUG)
+            .setExecutor(SynchronousExecutor())
+            .build()
+        WorkManagerTestInitHelper.initializeTestWorkManager(context, config)
+    }
+
     @Test
     @Throws(Exception::class)
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
+    fun testRecentWork() {
 
-        assertEquals("knf.kuma", appContext.packageName)
     }
 }
