@@ -1,6 +1,5 @@
 package knf.kuma.animeinfo
 
-import android.content.ActivityNotFoundException
 import android.content.Context
 import android.content.Intent
 import android.content.pm.ActivityInfo
@@ -26,6 +25,7 @@ import com.simplecityapps.recyclerview_fastscroll.views.FastScrollRecyclerView
 import com.squareup.picasso.Callback
 import knf.kuma.App
 import knf.kuma.R
+import knf.kuma.ads.AdsUtils
 import knf.kuma.animeinfo.fragments.ChaptersFragmentMaterial
 import knf.kuma.animeinfo.ktx.epTitle
 import knf.kuma.animeinfo.ktx.fileName
@@ -45,6 +45,7 @@ import knf.kuma.videoservers.ServersFactory
 import kotlinx.coroutines.*
 import org.jetbrains.anko.doAsync
 import xdroid.toaster.Toaster
+import java.net.URL
 import java.util.*
 import java.util.concurrent.atomic.AtomicReference
 
@@ -349,9 +350,29 @@ class AnimeChaptersAdapterMaterial(private val fragment: Fragment, private val r
                             R.id.commentaries -> {
                                 fragment.lifecycleScope.launch(Dispatchers.Main){
                                     try {
-                                        CommentariesDialog.show(fragment, withContext(Dispatchers.IO){ chapter.chapter.commentariesLink() })
-                                    } catch (e: ActivityNotFoundException) {
-                                        noCrashSuspend { context.startActivity(Intent(Intent.ACTION_VIEW, Uri.parse(withContext(Dispatchers.IO){ chapter.chapter.commentariesLink() }))) }
+                                        val version = withContext(Dispatchers.IO) {
+                                            Regex("load\\.(\\w+)\\.js").find(URL("https://https-animeflv-net.disqus.com/embed.js").readText())?.destructured?.component1()
+                                        }
+                                        CommentariesDialog.show(
+                                            fragment,
+                                            withContext(Dispatchers.IO) {
+                                                chapter.chapter.commentariesLink(version)
+                                            })
+                                    } catch (e: Exception) {
+                                        noCrashSuspend {
+                                            context.startActivity(
+                                                Intent(
+                                                    Intent.ACTION_VIEW,
+                                                    Uri.parse(withContext(Dispatchers.IO) {
+                                                        chapter.chapter.commentariesLink(
+                                                            AdsUtils.remoteConfigs.getString(
+                                                                "disqus_version"
+                                                            )
+                                                        )
+                                                    })
+                                                )
+                                            )
+                                        }
                                     }
                                 }
                             }
