@@ -11,7 +11,6 @@ import android.util.Log
 import androidx.core.app.NotificationCompat
 import androidx.core.content.ContextCompat
 import androidx.core.content.FileProvider
-import androidx.lifecycle.Observer
 import androidx.preference.PreferenceManager
 import androidx.tvprovider.media.tv.PreviewChannelHelper
 import androidx.work.*
@@ -29,12 +28,15 @@ import knf.kuma.pojos.Recents
 import knf.kuma.recents.RecentsNotReceiver
 import knf.kuma.search.SearchAdvObject
 import knf.kuma.tv.ChannelUtils
-import org.jetbrains.anko.doAsync
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 import org.jetbrains.anko.notificationManager
 import pl.droidsonroids.jspoon.Jspoon
 import java.util.concurrent.TimeUnit
+import kotlin.random.Random
 
-class RecentsWork(val context: Context, workerParameters: WorkerParameters) : Worker(context, workerParameters) {
+class RecentsWork(val context: Context, workerParameters: WorkerParameters) :
+    CoroutineWorker(context, workerParameters) {
     private val RECENTS_GROUP = "recents-group"
     private val recentsDAO = CacheDB.INSTANCE.recentsDAO()
     private val favsDAO = CacheDB.INSTANCE.favsDAO()
@@ -46,10 +48,14 @@ class RecentsWork(val context: Context, workerParameters: WorkerParameters) : Wo
     private val summaryBroadcast: Intent
         get() = Intent(context, RecentsNotReceiver::class.java).putExtra("mode", 1)
 
-    override fun doWork(): Result {
+    override suspend fun doWork(): Result {
         if (!Network.isConnected) return Result.success().also { Log.e("Recents", "No Network") }
+        //setForeground(createForegroundInfo())
         try {
-            val recents = Jspoon.create().adapter(Recents::class.java).fromHtml(jsoupCookies("https://animeflv.net/").get().outerHtml())
+            val recents = withContext(Dispatchers.IO) {
+                Jspoon.create().adapter(Recents::class.java)
+                    .fromHtml(jsoupCookies("https://animeflv.net/").get().outerHtml())
+            }
             val objects = RecentObject.create(recents.list ?: listOf())
             for ((i, recentObject) in objects.withIndex())
                 recentObject.key = i
@@ -70,8 +76,33 @@ class RecentsWork(val context: Context, workerParameters: WorkerParameters) : Wo
         }
     }
 
+    private fun createForegroundInfo(): ForegroundInfo =
+        ForegroundInfo(
+            Random.nextInt(1000, 9999),
+            NotificationCompat.Builder(context, CHANNEL_RECENTS)
+                .setSmallIcon(R.drawable.ic_recents_group)
+                .setColor(ContextCompat.getColor(context, R.color.colorAccent))
+                .setContentText("Buscando nuevos episodios")
+                .setProgress(100, 0, true)
+                .build()
+        )
+
+    private fun notifyTest() {
+        manager.notify(
+            Random.nextInt(),
+            NotificationCompat.Builder(context, CHANNEL_RECENTS)
+                .setSmallIcon(R.drawable.ic_recents_group)
+                .setColor(ContextCompat.getColor(context, R.color.colorAccent))
+                .setContentText("Test notification, ${System.currentTimeMillis()}")
+                .build()
+        )
+    }
+
     @Throws(Exception::class)
-    private fun notifyAllChaps(local: MutableList<RecentObject>, objects: MutableList<RecentObject>) {
+    private fun notifyAllChaps(
+        local: MutableList<RecentObject>,
+        objects: MutableList<RecentObject>
+    ) {
         for (recentObject in objects) {
             if (!local.contains(recentObject)) {
                 notifyRecent(recentObject)
@@ -129,18 +160,40 @@ class RecentsWork(val context: Context, workerParameters: WorkerParameters) : Wo
             val tone = FileAccessHelper.toneFile
             if (tone.exists())
                 setSound(
-                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                            val uri: Uri = FileProvider.getUriForFile(context, "${context.packageName}.fileprovider", tone)
-                            context.grantUriPermission("com.android.systemui", uri, Intent.FLAG_GRANT_READ_URI_PERMISSION)
-                            uri
-                        } else
-                            Uri.fromFile(tone)
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                        val uri: Uri = FileProvider.getUriForFile(
+                            context,
+                            "${context.packageName}.fileprovider",
+                            tone
+                        )
+                        context.grantUriPermission(
+                            "com.android.systemui",
+                            uri,
+                            Intent.FLAG_GRANT_READ_URI_PERMISSION
+                        )
+                        uri
+                    } else
+                        Uri.fromFile(tone)
                 )
             setLargeIcon(getBitmap(recentObject))
             setAutoCancel(true)
             setOnlyAlertOnce(true)
-            setContentIntent(PendingIntent.getActivity(context, System.currentTimeMillis().toInt(), getAnimeIntent(animeObject, obj), PendingIntent.FLAG_UPDATE_CURRENT))
-            setDeleteIntent(PendingIntent.getBroadcast(context, System.currentTimeMillis().toInt(), obj.getBroadcast(context), PendingIntent.FLAG_UPDATE_CURRENT))
+            setContentIntent(
+                PendingIntent.getActivity(
+                    context,
+                    System.currentTimeMillis().toInt(),
+                    getAnimeIntent(animeObject, obj),
+                    PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
+                )
+            )
+            setDeleteIntent(
+                PendingIntent.getBroadcast(
+                    context,
+                    System.currentTimeMillis().toInt(),
+                    obj.getBroadcast(context),
+                    PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
+                )
+            )
             if (isFullMode && !PrefsUtil.isFamilyFriendly)
                 addAction(
                     android.R.drawable.stat_sys_download_done,
@@ -149,7 +202,7 @@ class RecentsWork(val context: Context, workerParameters: WorkerParameters) : Wo
                         context,
                         System.currentTimeMillis().toInt(),
                         getChapIntent(recentObject, obj),
-                        PendingIntent.FLAG_UPDATE_CURRENT
+                        PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
                     )
                 )
             setGroup(RECENTS_GROUP)
@@ -207,15 +260,29 @@ class RecentsWork(val context: Context, workerParameters: WorkerParameters) : Wo
 
     private fun notifySummary() {
         val notification = NotificationCompat.Builder(context, CHANNEL_RECENTS)
-                .setSmallIcon(R.drawable.ic_recents_group)
-                .setColor(ContextCompat.getColor(context, R.color.colorAccent))
-                .setContentTitle("Nuevos capitulos")
-                .setContentText("Hay nuevos capitulos recientes!!")
-                .setGroupSummary(true)
-                .setGroup(RECENTS_GROUP)
-                .setAutoCancel(true)
-                .setContentIntent(PendingIntent.getActivity(context, 0, Intent(context, DesignUtils.mainClass), PendingIntent.FLAG_CANCEL_CURRENT))
-                .setDeleteIntent(PendingIntent.getBroadcast(context, System.currentTimeMillis().toInt(), summaryBroadcast, PendingIntent.FLAG_UPDATE_CURRENT))
+            .setSmallIcon(R.drawable.ic_recents_group)
+            .setColor(ContextCompat.getColor(context, R.color.colorAccent))
+            .setContentTitle("Nuevos capitulos")
+            .setContentText("Hay nuevos capitulos recientes!!")
+            .setGroupSummary(true)
+            .setGroup(RECENTS_GROUP)
+            .setAutoCancel(true)
+            .setContentIntent(
+                PendingIntent.getActivity(
+                    context,
+                    0,
+                    Intent(context, DesignUtils.mainClass),
+                    PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_IMMUTABLE
+                )
+            )
+            .setDeleteIntent(
+                PendingIntent.getBroadcast(
+                    context,
+                    System.currentTimeMillis().toInt(),
+                    summaryBroadcast,
+                    PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
+                )
+            )
                 .build()
         if (PrefsUtil.isGroupingEnabled)
             manager.notify(KEY_SUMMARY, notification)
@@ -226,8 +293,21 @@ class RecentsWork(val context: Context, workerParameters: WorkerParameters) : Wo
         const val KEY_SUMMARY = 55971
         internal const val TAG = "recents-job"
 
-        fun schedule(context: Context) {
-            WorkManager.getInstance(context).getWorkInfosByTagLiveData(TAG).let { ld ->
+        suspend fun schedule(context: Context) {
+            val preferences = PreferenceManager.getDefaultSharedPreferences(context)
+            val time = (preferences.getString("recents_time", "1") ?: "1").toInt() * 15
+            PeriodicWorkRequestBuilder<RecentsWork>(
+                time.coerceAtLeast(15).toLong(),
+                TimeUnit.MINUTES,
+                5,
+                TimeUnit.MINUTES
+            ).apply {
+                setInitialDelay(15L, TimeUnit.MINUTES)
+                //setConstraints(networkConnectedConstraints())
+                //setBackoffCriteria(BackoffPolicy.EXPONENTIAL, 1, TimeUnit.MINUTES)
+                addTag(TAG)
+            }.build().enqueueUnique(TAG, ExistingPeriodicWorkPolicy.KEEP)
+            /*WorkManager.getInstance(context).getWorkInfosByTagLiveData(TAG).let { ld ->
                 lateinit var observer: Observer<List<WorkInfo>>
                 doOnUI {
                     ld.observeForever(Observer<List<WorkInfo>> {
@@ -238,23 +318,27 @@ class RecentsWork(val context: Context, workerParameters: WorkerParameters) : Wo
                                 val time = (preferences.getString("recents_time", "1")
                                         ?: "1").toInt() * 15
                                 if (time > 0)
-                                    PeriodicWorkRequestBuilder<RecentsWork>(time.toLong(), TimeUnit.MINUTES, 5, TimeUnit.MINUTES).apply {
+                                    PeriodicWorkRequestBuilder<RecentsWork>(time.coerceAtLeast(15).toLong(), TimeUnit.MINUTES).apply {
+                                        setInitialDelay(15L, TimeUnit.MINUTES)
                                         //setConstraints(networkConnectedConstraints())
-                                        setBackoffCriteria(BackoffPolicy.EXPONENTIAL, 1, TimeUnit.MINUTES)
+                                        //setBackoffCriteria(BackoffPolicy.EXPONENTIAL, 1, TimeUnit.MINUTES)
                                         addTag(TAG)
                                     }.build().enqueueUnique(TAG, ExistingPeriodicWorkPolicy.REPLACE)
                             }
                     }.also { observer = it })
                 }
-            }
+            }*/
         }
 
         fun reSchedule(time: Int) {
             WorkManager.getInstance(App.context).cancelAllWorkByTag(TAG)
             if (time > 0)
-                PeriodicWorkRequestBuilder<RecentsWork>(time.toLong(), TimeUnit.MINUTES, 5, TimeUnit.MINUTES).apply {
+                PeriodicWorkRequestBuilder<RecentsWork>(
+                    time.coerceAtLeast(15).toLong(),
+                    TimeUnit.MINUTES
+                ).apply {
                     //setConstraints(networkConnectedConstraints())
-                    setBackoffCriteria(BackoffPolicy.EXPONENTIAL, 1, TimeUnit.MINUTES)
+                    //setBackoffCriteria(BackoffPolicy.EXPONENTIAL, 1, TimeUnit.MINUTES)
                     addTag(TAG)
                 }.build().enqueueUnique(TAG, ExistingPeriodicWorkPolicy.REPLACE)
         }
