@@ -71,17 +71,21 @@ class ConfigurationFragmentMaterial : PreferenceFragmentCompat() {
         super.onAttach(activity)
     }
 
-    @ExperimentalContracts
+
+    @OptIn(ExperimentalContracts::class)
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
         if (activity != null && context != null)
             doOnUI {
                 addPreferencesFromResource(R.xml.preferences)
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N)
-                    preferenceScreen.findPreference<Preference>(keyCustomTone)?.summary = "Abrir configuración"
+                    preferenceScreen.findPreference<Preference>(keyCustomTone)?.summary =
+                        "Abrir configuración"
                 else if (FileAccessHelper.toneFile.exists())
-                    preferenceScreen.findPreference<Preference>(keyCustomTone)?.summary = "Personalizado"
+                    preferenceScreen.findPreference<Preference>(keyCustomTone)?.summary =
+                        "Personalizado"
                 if (!DesignUtils.isFlat)
-                    preferenceScreen.findPreference<ListPreference>("recentActionType")?.isVisible = false
+                    preferenceScreen.findPreference<ListPreference>("recentActionType")?.isVisible =
+                        false
                 preferenceScreen.findPreference<Preference>(keyCustomTone)?.onPreferenceClickListener = Preference.OnPreferenceClickListener {
                     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
                         noCrash {
@@ -121,16 +125,29 @@ class ConfigurationFragmentMaterial : PreferenceFragmentCompat() {
                                         val autoBackupObject = it as? AutoBackupObject
                                         if (autoBackupObject != null) {
                                             if (autoBackupObject == AutoBackupObject(activity))
-                                                preferenceScreen.findPreference<Preference>(keyAutoBackup)?.summary = "%s"
+                                                preferenceScreen.findPreference<Preference>(
+                                                    keyAutoBackup
+                                                )?.summary = "%s"
                                             else
-                                                preferenceScreen.findPreference<Preference>(keyAutoBackup)?.summary = "Solo " + autoBackupObject.name
+                                                preferenceScreen.findPreference<Preference>(
+                                                    keyAutoBackup
+                                                )?.summary = "Solo " + autoBackupObject.name
                                             if (autoBackupObject.value == null)
                                                 GlobalScope.launch(Dispatchers.Main) {
-                                                    Backups.createService()?.backup(AutoBackupObject(App.context), Backups.keyAutoBackup)
-                                                    preferenceScreen.findPreference<Preference>(keyAutoBackup)?.summary = "%s"
+                                                    Backups.createService()?.backup(
+                                                        AutoBackupObject(App.context),
+                                                        Backups.keyAutoBackup
+                                                    )
+                                                    preferenceScreen.findPreference<Preference>(
+                                                        keyAutoBackup
+                                                    )?.summary = "%s"
                                                 }
                                             else
-                                                preferenceManager.sharedPreferences.edit().putString(keyAutoBackup, autoBackupObject.value).apply()
+                                                preferenceManager.sharedPreferences?.edit()
+                                                    ?.putString(
+                                                        keyAutoBackup,
+                                                        autoBackupObject.value
+                                                    )?.apply()
                                         } else {
                                             preferenceScreen.findPreference<Preference>(keyAutoBackup)?.summary = "%s (NE)"
                                         }
@@ -172,10 +189,15 @@ class ConfigurationFragmentMaterial : PreferenceFragmentCompat() {
                 } else {
                     preferenceScreen.removePreferenceRecursively("download_type_q")
                     preferenceScreen.findPreference<Preference>("download_type")?.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { _, newValue ->
-                        if (newValue == "1" && !FileAccessHelper.canDownload(this@ConfigurationFragmentMaterial, newValue as? String))
+                        if (newValue == "1" && !FileAccessHelper.canDownload(
+                                this@ConfigurationFragmentMaterial,
+                                newValue as? String
+                            )
+                        )
                             Toaster.toast("Por favor selecciona la raiz de tu SD")
                         else
-                            PreferenceManager.getDefaultSharedPreferences(context).edit().putString("tree_uri", null).apply()
+                            PreferenceManager.getDefaultSharedPreferences(requireContext()).edit()
+                                .putString("tree_uri", null).apply()
                         true
                     }
                 }
@@ -345,7 +367,7 @@ class ConfigurationFragmentMaterial : PreferenceFragmentCompat() {
                     false
                 }
                 if (!canGroupNotifications)
-                    preferenceScreen.removePreference(preferenceScreen.findPreference("group_notifications"))
+                    preferenceScreen.removePreference(preferenceScreen.findPreference("group_notifications")!!)
                 preferenceScreen.findPreference<Preference>("dir_destroy")?.onPreferenceClickListener = Preference.OnPreferenceClickListener {
                     try {
                         if (!DirectoryUpdateService.isRunning && !DirectoryService.isRunning)
@@ -379,9 +401,10 @@ class ConfigurationFragmentMaterial : PreferenceFragmentCompat() {
                             true
                         }
                     0 -> {
-                        val category = preferenceScreen.findPreference("category_design") as? PreferenceCategory
-                        category?.removePreference(preferenceScreen.findPreference(keyThemeColor))
-                        val pref = Preference(activity)
+                        val category =
+                            preferenceScreen.findPreference("category_design") as? PreferenceCategory
+                        category?.removePreference(preferenceScreen.findPreference(keyThemeColor)!!)
+                        val pref = Preference(requireContext())
                         pref.title = "Color de tema"
                         pref.summary = "Resuelve el secreto para desbloquear"
                         pref.setIcon(R.drawable.ic_palette)
