@@ -49,7 +49,8 @@ class TopActivity : GenericActivity() {
         }()
     }
 
-    @ExperimentalContracts
+
+    @OptIn(ExperimentalContracts::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         setTheme(EAHelper.getTheme())
         super.onCreate(savedInstanceState)
